feat: 댓글 삭제 시 검증 로직 추가 (#38)

+++ instragram/src/main/java/toyproject/instragram/comment/controller/CommentApiController.java
@@ -6,6 +6,8 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.web.bind.annotation.*;
 import toyproject.instragram.comment.controller.dto.CommentResponse;
+import toyproject.instragram.common.auth.SignIn;
+import toyproject.instragram.common.auth.SignInMember;
 import toyproject.instragram.common.response.CommonSliceResponse;
 import toyproject.instragram.common.response.SliceInfo;
 import toyproject.instragram.comment.entity.Comment;
@@ -15,6 +17,8 @@ import toyproject.instragram.reply.service.ReplyService;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static toyproject.instragram.common.exception.api.ApiExceptionType.FORBIDDEN_COMMENT;
+
 @RestController
 @RequiredArgsConstructor
 public class CommentApiController {
@@ -23,11 +27,11 @@ public class CommentApiController {
     private final ReplyService replyService;
 
     @GetMapping("/comments")
-    public CommonSliceResponse<CommentResponse> getComments(
+    public CommonSliceResponse<List<CommentResponse>> getComments(
             @RequestParam Long postId, @PageableDefault(size = 20) Pageable pageable) {
         Slice<Comment> commentSlice = commentService.getCommentSlice(postId, pageable);
 
-        return new CommonSliceResponse(getCommentResponses(commentSlice.getContent()), SliceInfo.from(commentSlice));
+        return new CommonSliceResponse<>(getCommentResponses(commentSlice.getContent()), SliceInfo.from(commentSlice));
     }
 
     private List<CommentResponse> getCommentResponses(List<Comment> contents) {
@@ -37,7 +41,19 @@ public class CommentApiController {
     }
 
     @DeleteMapping("/comments/{commentId}")
-    public void deleteComment(@PathVariable Long commentId) {
+    public void deleteComment(@SignIn SignInMember signInMember, @PathVariable Long commentId) {
+        validateCommentAccess(signInMember.getMemberId(), getCommentOwnerId(commentId));
         commentService.deleteComment(commentId);
     }
+
+    private void validateCommentAccess(Long signInMemberId, Long commentOwnerId) {
+        if (!signInMemberId.equals(commentOwnerId)) {
+            throw FORBIDDEN_COMMENT.getException();
+        }
+    }
+
+    private Long getCommentOwnerId(Long commentId) {
+        return commentService.getComment(commentId).getMember().getId();
+    }
+
 }
