+++ instragram/src/main/java/toyproject/instragram/comment/service/CommentService.java
@@ -8,8 +8,11 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import toyproject.instragram.comment.entity.Comment;
 import toyproject.instragram.comment.repository.CommentRepository;
+import toyproject.instragram.common.exception.api.ApiExceptionType;
 import toyproject.instragram.member.repository.MemberRepository;
 import toyproject.instragram.post.repository.PostRepository;
+
+import static toyproject.instragram.common.exception.api.ApiExceptionType.*;
 
 @Service
 @Transactional(readOnly = true)
@@ -19,8 +22,6 @@ public class CommentService {
     private final PostRepository postRepository;
     private final MemberRepository memberRepository;
     private final CommentRepository commentRepository;
-
-    private final int MAX_COMMENT_SIZE = 20;
 
     @Transactional
     public Long addComment(CommentDto commentDto) {
@@ -39,11 +40,21 @@ public class CommentService {
 
     @Transactional
     public void deleteComment(Long commentId) {
+        if (!commentRepository.existsById(commentId)) {
+            throw NOT_FOUND_COMMENT.getException();
+        }
+
         commentRepository.deleteById(commentId);
     }
 
     // TODO 테스트 코드 작성
     public Long getCommentCount(Long postId) {
         return commentRepository.countCommentsByPostId(postId);
+    }
+
+    // TODO 테스트 코드 작성
+    public Comment getComment(Long commentId) {
+        return commentRepository.findById(commentId)
+                .orElseThrow(NOT_FOUND_COMMENT::getException);
     }
 }
